refactor files acc to samples standard
