--- vertexai/snippets/src/main/java/vertexai/gemini/samples/QuestionAnswer.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package vertexai.gemini.samples;
-
-import com.google.cloud.vertexai.VertexAI;
-import com.google.cloud.vertexai.generativeai.preview.GenerativeModel;
-import com.google.cloud.vertexai.generativeai.preview.ResponseHandler;
-import com.google.cloud.vertexai.api.GenerateContentResponse;
-
-public class QuestionAnswer {
-
-    private static final String PROJECT_ID = System.getenv("GOOGLE_CLOUD_PROJECT");
-    private static final String LOCATION = "us-central1";
-    private static final String MODEL_NAME = "gemini-pro-vision";
-
-    /**
-     * Ask a simple question to the model.
-     *
-     * @return the answer from the model
-     * @throws Exception if connection to the model fails
-     */
-    public static String simpleQuestion() throws Exception {
-        String output;
-
-        // Initialize client that will be used to send requests.
-        // This client only needs to be created once, and can be reused for multiple requests.
-        try (VertexAI vertexAI = new VertexAI(PROJECT_ID, LOCATION)) {
-            GenerativeModel model = new GenerativeModel(MODEL_NAME, vertexAI);
-
-            GenerateContentResponse response = model.generateContent("Why is the sky blue?");
-            output = ResponseHandler.getText(response);
-        }
-
-        return output;
-    }
-}
